Fix the issue of filename start with same 'blah.' will make some of them disappear
Fix the issue:

When including files which start with same 'blah.',
such as: jquery.validate.js and jquery.cookies.js (both of them are
start with 'jquery.'), after building project, some of them will disappear.

I think this fix method is stupid (but it works),
so please kindly provide the better way.